\name{Ss2.MTA.gene.annot}
\alias{Ss2.MTA.gene.annot}
\non_function{}
\title{Ss2.MTA.gene.annot}
\description{\code{data.table} environment describing Affymetrix/Thermo-Fisher annotations (from release 36) for the probe grouping method (gene-level or exon-level) for use in the \code{Sscore2} package }
\keyword{datasets}
