\name{Ss2.MTA.gene.bgplist}
\alias{Ss2.MTA.gene.bgplist}
\non_function{}
\title{Ss2.MTA.gene.bgplist}
\description{\code{data.table} environment describing Affymetrix/Thermo-Fisher id names and the number of probes for the 16,943 antigemonic background probes for use in the \code{Sscore2} package }
\keyword{datasets}
